Fix a bug in repo_list_page() that manifests only on 32-bit builds on big-endian CPUs.
FossilOrigin-Name: 42076cabfa45198898ebaba17b044bdb038e6c8e5e1ff56f0ce12b2e57d5f32f

+++ src/repolist.c
@@ -285,7 +285,7 @@ int repo_list_page(void){
       }
       blob_append_sql(&html,
         "<td></td><td data-sortkey='%08x'>%h</td>\n",
-        iAge, zAge);
+        (int)iAge, zAge);
       fossil_free(zAge);
       if( x.zLoginGroup ){
         blob_append_sql(&html, "<td></td><td>%h</td></tr>\n", x.zLoginGroup);
